+++ Corona-Warn-App/src/deviceForTesters/java/de/rki/coronawarnapp/test/presencetracing/ui/PresenceTracingTestViewModel.kt
@@ -61,7 +61,7 @@ class PresenceTracingTestViewModel @AssistedInject constructor(
             taskController.submitBlocking(
                 DefaultTaskRequest(
                     PresenceTracingWarningTask::class,
-                    originTag = "EventRegistrationTestFragmentViewModel"
+                    originTag = "PresenceTracingTestViewModel"
                 )
             )
         }
