+++ Corona-Warn-App/src/deviceForTesters/java/de/rki/coronawarnapp/ui/main/MainActivityTestModule.kt
@@ -71,8 +71,8 @@ abstract class MainActivityTestModule {
     abstract fun deltaOnboarding(): DeltaonboardingFragment
 
     @ContributesAndroidInjector(modules = [PresenceTracingTestFragmentModule::class])
-    abstract fun eventRegistration(): PresenceTracingTestFragment
+    abstract fun presenceTracingTestFragment(): PresenceTracingTestFragment
 
     @ContributesAndroidInjector(modules = [QrCodePosterTestFragmentModule::class])
-    abstract fun showEventDetail(): QrCodePosterTestFragment
+    abstract fun qrCodePosterTestFragment(): QrCodePosterTestFragment
 }
