+++ Corona-Warn-App/src/main/res/navigation/nav_graph.xml
@@ -9,7 +9,7 @@
     <include app:graph="@navigation/test_nav_graph" />
     <!--    Contact Diary graph-->
     <include app:graph="@navigation/contact_diary_nav_graph" />
-    <!--    Event registration graphs-->
+    <!--    Presence Tracing graphs-->
     <include app:graph="@navigation/trace_location_organizer_nav_graph" />
     <include app:graph="@navigation/trace_location_attendee_nav_graph" />
