+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/presencetracing/checkins/CheckInRetentionTest.kt
@@ -1,8 +1,5 @@
 package de.rki.coronawarnapp.presencetracing.checkins
 
-import de.rki.coronawarnapp.presencetracing.checkins.CheckIn
-import de.rki.coronawarnapp.presencetracing.checkins.isOutOfRetention
-import de.rki.coronawarnapp.presencetracing.checkins.isWithinRetention
 import io.kotest.matchers.shouldBe
 import okio.ByteString.Companion.encode
 import org.joda.time.Instant
