+++ Corona-Warn-App/src/test/java/de/rki/coronawarnapp/presencetracing/checkins/CheckInTransmissionRiskLevelTest.kt
@@ -1,7 +1,5 @@
 package de.rki.coronawarnapp.presencetracing.checkins
 
-import de.rki.coronawarnapp.presencetracing.checkins.CheckIn
-import de.rki.coronawarnapp.presencetracing.checkins.determineRiskTransmission
 import de.rki.coronawarnapp.submission.task.TransmissionRiskVector
 import io.kotest.matchers.shouldBe
 import okio.ByteString.Companion.encode
